Split and-then to two sections and added exercise.
